backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
anthropic==0.52.0
python-dotenv==1.1.0
pyyaml==6.0.2
sentence-transformers==2.7.0
numpy==1.26.4
python-multipart==0.0.9
